OtelContext moved to telemetry subpackage
